+++ ChangeLog.md
@@ -49,6 +49,7 @@ Sat, `30th October 2016`
         - Updated heading
 - Small cleaning
     - `active.php` header cleaned up
+    - `regular.php` header cleaned up
     
 
 ##### Changes